refactor(perf): split src/config/schemas.ts into 6 domain sub-files
