+++ src/config/schemas.ts
@@ -2,860 +2,41 @@
  * Zod Schema Definitions
  *
  * Runtime validation schemas for nax configuration.
+ * Sub-schemas are extracted into schemas-*.ts files.
  */
 
 import { z } from "zod";
+import { ContextConfigSchema } from "./schemas-context";
+import { DebateConfigSchema } from "./schemas-debate";
+import {
+  AutoModeConfigSchema,
+  ConstitutionConfigSchema,
+  ExecutionConfigSchema,
+  QualityConfigSchema,
+  TddConfigSchema,
+} from "./schemas-execution";
+import {
+  AcceptanceConfigSchema,
+  AgentConfigSchema,
+  GenerateConfigSchema,
+  HooksConfigSchema,
+  InteractionConfigSchema,
+  OptimizerConfigSchema,
+  PlanConfigSchema,
+  PluginConfigEntrySchema,
+  PrecheckConfigSchema,
+  ProjectProfileSchema,
+  PromptsConfigSchema,
+  RoutingConfigSchema,
+} from "./schemas-infra";
+import { ModelMapSchema } from "./schemas-model";
+import { ReviewConfigSchema } from "./schemas-review";
 
-/** Zod schema for runtime validation */
-const TokenPricingSchema = z.object({
-  inputPer1M: z.number().min(0),
-  outputPer1M: z.number().min(0),
-});
-
-const ModelDefSchema = z.object({
-  provider: z.string().min(1, "Provider must be non-empty"),
-  model: z.string().min(1, "Model must be non-empty"),
-  pricing: TokenPricingSchema.optional(),
-  env: z.record(z.string(), z.string()).optional(),
-});
-
-const ModelEntrySchema = z.union([z.string().min(1, "Model identifier must be non-empty"), ModelDefSchema]);
-
-/** Detect legacy flat format: any top-level value is a string or has 'provider'/'model' key directly */
-function isLegacyFlatModels(val: unknown): boolean {
-  if (typeof val !== "object" || val === null) return false;
-  const obj = val as Record<string, unknown>;
-  for (const v of Object.values(obj)) {
-    if (typeof v === "string") return true;
-    if (typeof v === "object" && v !== null && ("provider" in v || "model" in v)) return true;
-  }
-  return false;
-}
-
-/** Per-agent model map: Record<agentName, Record<tierName, ModelEntry>> */
-const PerAgentModelMapSchema = z.record(z.string().min(1), z.record(z.string().min(1), ModelEntrySchema));
-
-const ModelMapSchema = z.preprocess((val) => {
-  if (isLegacyFlatModels(val)) {
-    return { claude: val };
-  }
-  return val;
-}, PerAgentModelMapSchema);
-
-const ModelTierSchema = z.string().min(1, "Tier name must be non-empty");
-const ConfiguredModelObjectSchema = z.object({
-  agent: z.string().min(1, "agent must be non-empty"),
-  model: z.string().min(1, "model must be non-empty"),
-});
-const ConfiguredModelSchema = z.union([ModelTierSchema, ConfiguredModelObjectSchema]);
-
-const TierConfigSchema = z.object({
-  tier: z.string().min(1, "Tier name must be non-empty"),
-  attempts: z.number().int().min(1).max(20, { message: "attempts must be 1-20" }),
-});
-
-const AutoModeConfigSchema = z.object({
-  enabled: z.boolean(),
-  complexityRouting: z.object({
-    simple: ModelTierSchema,
-    medium: ModelTierSchema,
-    complex: ModelTierSchema,
-    expert: ModelTierSchema,
-  }),
-  escalation: z.object({
-    enabled: z.boolean(),
-    tierOrder: z.array(TierConfigSchema).min(1, { message: "tierOrder must have at least one tier" }),
-    escalateEntireBatch: z.boolean().optional(),
-  }),
-});
-
-const RectificationConfigSchema = z.object({
-  enabled: z.boolean().default(true),
-  maxRetries: z.number().int().min(0).max(10).default(2),
-  fullSuiteTimeoutSeconds: z.number().int().min(10).max(600).default(120),
-  maxFailureSummaryChars: z.number().int().min(500).max(10000).default(2000),
-  abortOnIncreasingFailures: z.boolean().default(true),
-  escalateOnExhaustion: z.boolean().optional().default(true),
-  rethinkAtAttempt: z.number().int().min(1).default(2),
-  urgencyAtAttempt: z.number().int().min(1).default(3),
-});
-
-const RegressionGateConfigSchema = z.object({
-  enabled: z.boolean().default(true),
-  timeoutSeconds: z.number().int().min(10).max(600).default(120),
-  acceptOnTimeout: z.boolean().default(true),
-  mode: z.enum(["deferred", "per-story", "disabled"]).default("deferred"),
-  maxRectificationAttempts: z.number().int().min(1).default(2),
-});
-
-const SmartTestRunnerConfigSchema = z.object({
-  enabled: z.boolean().default(true),
-  /**
-   * Optional — undefined means user did not set this; resolver falls through
-   * to auto-detection then DEFAULT_TEST_FILE_PATTERNS. (ADR-009)
-   */
-  testFilePatterns: z.array(z.string()).optional(),
-  fallback: z.enum(["import-grep", "full-suite"]).default("import-grep"),
-});
-
-const SMART_TEST_RUNNER_DEFAULT = {
-  enabled: true,
-  fallback: "import-grep" as const,
-};
-
-/** Coerces boolean → SmartTestRunnerConfig for backward compat */
-const smartTestRunnerFieldSchema = z
-  .preprocess((val) => {
-    if (typeof val === "boolean") {
-      return { enabled: val, fallback: "import-grep" };
-    }
-    return val;
-  }, SmartTestRunnerConfigSchema)
-  .default(SMART_TEST_RUNNER_DEFAULT);
-
-const WorktreeDependenciesConfigSchema = z
-  .object({
-    mode: z.enum(["inherit", "provision", "off"]).default("off"),
-    setupCommand: z.string().nullable().default(null),
-  })
-  .superRefine((value, ctx) => {
-    if (value.mode !== "provision" && value.setupCommand !== null) {
-      ctx.addIssue({
-        code: z.ZodIssueCode.custom,
-        path: ["setupCommand"],
-        message: "execution.worktreeDependencies.setupCommand requires mode 'provision'",
-      });
-    }
-  });
-
-const ExecutionConfigSchema = z.object({
-  maxIterations: z.number().int().positive({ message: "maxIterations must be > 0" }),
-  iterationDelayMs: z.number().int().nonnegative(),
-  costLimit: z.number().positive({ message: "costLimit must be > 0" }),
-  sessionTimeoutSeconds: z.number().int().positive({ message: "sessionTimeoutSeconds must be > 0" }).default(3600),
-  /** Max retries when acpx signals a non-retryable session error (e.g. stale/locked session). */
-  sessionErrorMaxRetries: z.number().int().min(0).max(5).default(1),
-  /** Max retries when acpx signals a retryable session error (e.g. QUEUE_DISCONNECTED_BEFORE_COMPLETION). */
-  sessionErrorRetryableMaxRetries: z.number().int().min(0).max(10).default(3),
-  verificationTimeoutSeconds: z.number().int().min(1).max(3600).default(300),
-  maxStoriesPerFeature: z.number().int().positive(),
-  rectification: RectificationConfigSchema,
-  regressionGate: RegressionGateConfigSchema,
-  contextProviderTokenBudget: z
-    .number()
-    .int()
-    .positive({ message: "contextProviderTokenBudget must be > 0" })
-    .default(2000),
-  lintCommand: z.string().nullable().optional(),
-  typecheckCommand: z.string().nullable().optional(),
-  permissionProfile: z.enum(["unrestricted", "safe", "scoped"]).default("unrestricted"),
-  // Phase 2: per-stage permission overrides (only read when profile = "scoped")
-  permissions: z
-    .record(
-      z.string(),
-      z.object({
-        mode: z.enum(["approve-all", "approve-reads", "scoped"]),
-        allowedTools: z.array(z.string()).optional(),
-        inherit: z.string().optional(),
-      }),
-    )
-    .optional(),
-  smartTestRunner: smartTestRunnerFieldSchema,
-  worktreeDependencies: WorktreeDependenciesConfigSchema.default({
-    mode: "off",
-    setupCommand: null,
-  }),
-  storyIsolation: z.enum(["shared", "worktree"]).default("shared"),
-});
-
-const QualityConfigSchema = z.object({
-  requireTypecheck: z.boolean().default(true),
-  requireLint: z.boolean().default(true),
-  requireTests: z.boolean().default(true),
-  scopeTestThreshold: z.number().int().min(0).default(10),
-  commands: z
-    .object({
-      typecheck: z.string().optional(),
-      lint: z.string().optional(),
-      test: z.string().optional(),
-      testScoped: z.string().optional(),
-      lintFix: z.string().optional(),
-      formatFix: z.string().optional(),
-      build: z.string().optional(),
-    })
-    .default({}),
-  autofix: z
-    .object({
-      enabled: z.boolean().default(true),
-      maxAttempts: z.number().int().min(1).default(3),
-      maxTotalAttempts: z.number().int().min(1).default(12),
-      rethinkAtAttempt: z.number().int().min(1).default(2),
-      urgencyAtAttempt: z.number().int().min(1).default(3),
-    })
-    .default({
-      enabled: true,
-      maxAttempts: 3,
-      maxTotalAttempts: 12,
-      rethinkAtAttempt: 2,
-      urgencyAtAttempt: 3,
-    }),
-  forceExit: z.boolean().default(false),
-  detectOpenHandles: z.boolean().default(true),
-  detectOpenHandlesRetries: z.number().int().min(0).max(5).default(1),
-  gracePeriodMs: z.number().int().min(500).max(30000).default(5000),
-  drainTimeoutMs: z.number().int().min(0).max(10000).default(2000),
-  shell: z.string().default("/bin/sh"),
-  stripEnvVars: z
-    .array(z.string())
-    .default([
-      "CLAUDECODE",
-      "REPL_ID",
-      "AGENT",
-      "GITLAB_ACCESS_TOKEN",
-      "GITHUB_TOKEN",
-      "GITHUB_ACCESS_TOKEN",
-      "GH_TOKEN",
-      "CI_GIT_TOKEN",
-      "CI_JOB_TOKEN",
-      "BITBUCKET_ACCESS_TOKEN",
-      "NPM_TOKEN",
-      "NPM_AUTH_TOKEN",
-      "YARN_NPM_AUTH_TOKEN",
-      "ANTHROPIC_API_KEY",
-      "OPENAI_API_KEY",
-      "GEMINI_API_KEY",
-      "COHERE_API_KEY",
-      "AWS_ACCESS_KEY_ID",
-      "AWS_SECRET_ACCESS_KEY",
-      "AWS_SESSION_TOKEN",
-      "GOOGLE_APPLICATION_CREDENTIALS",
-      "GCLOUD_SERVICE_KEY",
-      "AZURE_CLIENT_SECRET",
-      "AZURE_TENANT_ID",
-      "TELEGRAM_BOT_TOKEN",
-      "SLACK_TOKEN",
-      "SLACK_WEBHOOK_URL",
-      "SENTRY_AUTH_TOKEN",
-      "DATADOG_API_KEY",
-    ]),
-
-  testing: z
-    .object({
-      /**
-       * When true (default), nax injects a hermetic test requirement into all code-writing prompts.
-       * Instructs the AI to mock all I/O boundaries (HTTP, CLI spawning, databases, etc.)
-       * and never invoke real external processes or services during test execution.
-       * Set to false only if your project requires real integration calls in tests.
-       */
-      hermetic: z.boolean().default(true),
-      /**
-       * Project-specific external boundaries the AI should watch for and mock.
-       * E.g. ["claude", "acpx", "redis", "grpc"] — any CLI tools, clients, or services
-       * the project uses that should never be called from tests.
-       */
-      externalBoundaries: z.array(z.string()).optional(),
-      /**
-       * Project-specific guidance on how to mock external dependencies.
-       * Injected verbatim into the hermetic requirement section of the prompt.
-       * E.g. "Use injectable deps for CLI spawning, ioredis-mock for Redis"
-       */
-      mockGuidance: z.string().optional(),
-    })
-    .optional(),
-});
-
-const TddConfigSchema = z.object({
-  maxRetries: z.number().int().nonnegative(),
-  autoVerifyIsolation: z.boolean(),
-  autoApproveVerifier: z.boolean(),
-  strategy: z.enum(["auto", "strict", "lite", "off"]).default("auto"),
-  sessionTiers: z
-    .object({
-      testWriter: z.string().optional(),
-      implementer: z.string().optional(),
-      verifier: z.string().optional(),
-    })
-    .optional(),
-  testWriterAllowedPaths: z.array(z.string()).optional(),
-  rollbackOnFailure: z.boolean().optional(),
-  greenfieldDetection: z.boolean().optional(),
-});
-
-const ConstitutionConfigSchema = z.object({
-  enabled: z.boolean(),
-  path: z.string().min(1, "constitution.path must be non-empty"),
-  maxTokens: z.number().int().positive({ message: "constitution.maxTokens must be > 0" }),
-  skipGlobal: z.boolean().optional(),
-});
-
-const SemanticReviewConfigSchema = z.object({
-  modelTier: ModelTierSchema.default("balanced"),
-  /**
-   * How the semantic reviewer accesses the git diff.
-   * "embedded": pre-collected diff truncated at 50KB and embedded in prompt.
-   * "ref" (default): only stat summary + storyGitRef passed; reviewer fetches full diff via tools.
-   */
-  diffMode: z.enum(["embedded", "ref"]).default("ref"),
-  /**
-   * When true, clears storyGitRef on failed stories during re-run initialization so
-   * the ref is re-captured at the next story start. Prevents cross-story diff pollution
-   * when multiple stories exhaust all tiers and are re-run. Default false (current behaviour).
-   */
-  resetRefOnRerun: z.boolean().default(false),
-  rules: z.array(z.string()).default([]),
-  timeoutMs: z.number().int().positive().default(600_000),
-  /**
-   * Optional — undefined means "derive from testFilePatterns + well-known noise dirs".
-   * Any user-set value (including []) is returned as-is. (ADR-009 §4.4)
-   */
-  excludePatterns: z.array(z.string()).optional(),
-});
-
-export const ReviewDialogueConfigSchema = z.object({
-  enabled: z.boolean().default(false),
-  maxClarificationsPerAttempt: z.number().int().min(0).max(10).default(2),
-  maxDialogueMessages: z.number().int().min(5).max(100).default(20),
-});
-
-/**
- * Adversarial review config — ships off by default (opt-in via review.checks).
- * Destructive heuristics: finds what is missing or broken, not what is present.
- */
-export const AdversarialReviewConfigSchema = z.object({
-  modelTier: ModelTierSchema.default("balanced"),
-  /**
-   * "ref" (default): reviewer self-serves the full diff via git tools — no 50KB cap,
-   *   test files included. Instructs reviewer to run git diff commands.
-   * "embedded": pre-collected full diff (no excludePatterns) embedded in prompt.
-   */
-  diffMode: z.enum(["embedded", "ref"]).default("ref"),
-  /** Custom adversarial heuristic rules to append to the prompt. */
-  rules: z.array(z.string()).default([]),
-  /** LLM call timeout in milliseconds. Default 600s (matches semantic — no debate path but ref mode may need full tool traversal). */
-  timeoutMs: z.number().int().positive().default(600_000),
-  /**
-   * Pathspec exclusions applied in embedded mode (to collectDiff) and in ref mode
-   * (shown in the prompt's git commands).
-   *
-   * Optional — undefined means "derive from testFilePatterns + noise dirs" (adversarial
-   * defaults to minimal exclusions so it sees test files). Any user-set value (including [])
-   * is returned as-is. (ADR-009 §4.4)
-   */
-  excludePatterns: z.array(z.string()).optional(),
-  /**
-   * When true, run semantic and adversarial reviewers concurrently via Promise.all.
-   * Default false (conservative rollout). Only activates when session count is within cap.
-   */
-  parallel: z.boolean().default(false),
-  /** Maximum combined reviewer sessions before falling back to sequential. Default 2. */
-  maxConcurrentSessions: z.number().int().min(1).max(4).default(2),
-});
-
-const ReviewConfigSchema = z.object({
-  enabled: z.boolean(),
-  checks: z.array(z.enum(["typecheck", "lint", "test", "build", "semantic", "adversarial"])),
-  commands: z.object({
-    typecheck: z.string().optional(),
-    lint: z.string().optional(),
-    test: z.string().optional(),
-    build: z.string().optional(),
-    lintFix: z.string().optional(),
-    formatFix: z.string().optional(),
-  }),
-  pluginMode: z.enum(["per-story", "deferred"]).default("per-story"),
-  audit: z.object({ enabled: z.boolean().default(false) }).default({ enabled: false }),
-  /**
-   * Minimum severity that counts as a blocking finding.
-   * "error"   (default): only error/critical findings block; warnings are advisory.
-   * "warning": error, critical, AND warning findings block; info is advisory.
-   * "info":    all findings block (strictest mode).
-   *
-   * Hierarchy: info < warning < error < critical.
-   * Applies only to LLM-based checkers (semantic, adversarial).
-   * Mechanical checks (lint, typecheck, test, build) always block on failure.
-   */
-  blockingThreshold: z.enum(["error", "warning", "info"]).default("error"),
-  semantic: SemanticReviewConfigSchema.optional(),
-  adversarial: AdversarialReviewConfigSchema.optional(),
-  dialogue: ReviewDialogueConfigSchema.default({
-    enabled: false,
-    maxClarificationsPerAttempt: 2,
-    maxDialogueMessages: 20,
-  }),
-});
-
-const PlanConfigSchema = z.object({
-  model: ConfiguredModelSchema,
-  outputPath: z.string().min(1, "plan.outputPath must be non-empty"),
-  timeoutSeconds: z.number().int().positive().default(600),
-  /** Override timeout for decompose calls in seconds. Defaults to plan.timeoutSeconds. */
-  decomposeTimeoutSeconds: z.number().int().min(30).max(1_800).optional(),
-});
-
-const AcceptanceFixConfigSchema = z.object({
-  diagnoseModel: ConfiguredModelSchema.default("fast"),
-  fixModel: ConfiguredModelSchema.default("balanced"),
-  strategy: z.enum(["diagnose-first", "implement-only"]).default("diagnose-first"),
-  maxRetries: z.number().int().nonnegative().default(2),
-});
-
-export const AcceptanceConfigSchema = z.object({
-  enabled: z.boolean(),
-  maxRetries: z.number().int().nonnegative(),
-  generateTests: z.boolean(),
-  testPath: z.string().min(1, "acceptance.testPath must be non-empty"),
-  command: z.string().optional(),
-  model: ConfiguredModelSchema.default("fast"),
-  refinement: z.boolean().default(true),
-  refinementConcurrency: z.number().int().min(1).max(10).default(3),
-  redGate: z.boolean().default(true),
-  testStrategy: z.enum(["unit", "component", "cli", "e2e", "snapshot"]).optional(),
-  testFramework: z.string().min(1, "acceptance.testFramework must be non-empty").optional(),
-  timeoutMs: z.number().int().min(30000).max(3600000).default(1800000),
-  fix: AcceptanceFixConfigSchema.optional().default({
-    diagnoseModel: "fast",
-    fixModel: "balanced",
-    strategy: "diagnose-first",
-    maxRetries: 2,
-  }),
-  suggestedTestPath: z.string().min(1).optional(),
-  hardening: z
-    .object({
-      enabled: z.boolean().default(true),
-    })
-    .optional()
-    .default({ enabled: true }),
-});
-
-const TestCoverageConfigSchema = z.object({
-  enabled: z.boolean().default(true),
-  detail: z.enum(["names-only", "names-and-counts", "describe-blocks"]).default("names-and-counts"),
-  maxTokens: z.number().int().min(50).max(5000).default(500),
-  testDir: z.string().optional(),
-  /** @deprecated Migrate to execution.smartTestRunner.testFilePatterns. Migration shim in src/config/migrations.ts. */
-  testPattern: z.string().optional(),
-  scopeToStory: z.boolean().default(true),
-});
-
-const ContextAutoDetectConfigSchema = z.object({
-  enabled: z.boolean().default(true),
-  maxFiles: z.number().int().min(1).max(20).default(5),
-  traceImports: z.boolean().default(false),
-});
-
-const FeatureContextEngineConfigSchema = z.object({
-  enabled: z.boolean().default(false),
-  budgetTokens: z.number().int().min(256).default(2048),
-});
-
-// Context Engine v2 pull tool config (Phase 4)
-const ContextV2PullConfigSchema = z
-  .object({
-    /**
-     * Enable pull tools for this run.
-     * When false (default), assemble() returns an empty pullTools array.
-     */
-    enabled: z.boolean().default(false),
-    /**
-     * Tool names permitted to activate. Empty array = all stage-configured tools allowed.
-     * Use to restrict which tools are enabled without changing the stage map.
-     */
-    allowedTools: z.array(z.string()).default([]),
-    /**
-     * Per-session call ceiling (overrides the descriptor default when set).
-     */
-    maxCallsPerSession: z.number().int().min(0).default(5),
-    /**
-     * Per-run call ceiling across all sessions in a single nax run.
-     */
-    maxCallsPerRun: z.number().int().min(0).default(50),
-  })
-  .default(() => ({ enabled: false, allowedTools: [], maxCallsPerSession: 5, maxCallsPerRun: 50 }));
-
-// Context Engine v2 rules config (Phase 5.1)
-const ContextV2RulesConfigSchema = z
-  .object({
-    /**
-     * Fall back to reading CLAUDE.md + .claude/rules/ when .nax/rules/ is absent.
-     * Default false — legacy fallback must be explicitly opted in.
-     * Set true only during migration to the canonical .nax/rules/ store.
-     */
-    allowLegacyClaudeMd: z.boolean().default(false),
-    /**
-     * Token ceiling for canonical rules. Lower-priority rules are tail-truncated
-     * when this budget is exceeded.
-     */
-    budgetTokens: z.number().int().min(512).default(8192),
-  })
-  .default(() => ({ allowLegacyClaudeMd: false, budgetTokens: 8192 }));
-
-// Context Engine plugin provider config (Phase 7)
-const ContextPluginProviderConfigSchema = z.object({
-  /**
-   * Module specifier for the plugin provider.
-   * Accepts npm package names (e.g. "@company/nax-rag") or paths
-   * relative to the project workdir (e.g. "./plugins/my-provider.js").
-   */
-  module: z.string().min(1),
-  /**
-   * Provider-specific config object passed to provider.init(config) on load.
-   * Shape is provider-defined — the engine passes it through opaquely.
-   */
-  config: z.record(z.string(), z.unknown()).optional(),
-  /**
-   * Set false to skip this provider without removing the config entry.
-   * Useful for temporarily disabling a provider for debugging.
-   */
-  enabled: z.boolean().default(true),
-});
-
-const ContextV2StageOverrideSchema = z.object({
-  budgetTokens: z.number().int().positive().optional(),
-  extraProviderIds: z.array(z.string().min(1)).default([]),
-});
-
-// Context Engine config (Phase 6: selective on; operators opt in per project)
-export const ContextV2ConfigSchema = z
-  .object({
-    /**
-     * Enable Context Engine orchestrator.
-     * Default: false — operators opt in by setting this true in their project config.
-     * Phase 6: selective on; Phase 7: plugin providers available once enabled.
-     */
-    enabled: z.boolean().default(false),
-    /**
-     * Minimum score threshold — chunks below this are dropped as noise.
-     * Phase 0: near-zero (0.1) so existing content is almost never filtered.
-     * Post-GA: tuned upward once effectiveness signal data is available.
-     */
-    minScore: z.number().min(0).max(1).default(0.1),
-    /** Pull tool configuration (Phase 4+) */
-    pull: ContextV2PullConfigSchema,
-    /** Canonical rules store configuration (Phase 5.1+) */
-    rules: ContextV2RulesConfigSchema,
-    /**
-     * External plugin provider registrations (Phase 7+).
-     * Each entry loads a module that exports an IContextProvider-compatible object.
-     * Empty by default — operators add providers for RAG, graph, or KB use cases.
-     */
-    pluginProviders: z.array(ContextPluginProviderConfigSchema).default([]),
-    /**
-     * Per-stage token budget overrides (AC-59).
-     * Keys are pipeline stage names (e.g. "execution", "tdd-implementer").
-     * Set in per-package config (<repoRoot>/.nax/mono/<packageDir>/config.json)
-     * to override the default stage budget for a specific package.
-     *
-     * Example: { "execution": { "budgetTokens": 15000 } }
-     */
-    stages: z.record(z.string().min(1), ContextV2StageOverrideSchema).default({}),
-    /**
-     * Determinism mode (AC-24).
-     * When true, providers that declare `deterministic: false` are excluded from assembly.
-     * Guarantees two runs with identical inputs produce identical push blocks.
-     * Default: false — all providers run regardless of determinism declaration.
-     */
-    deterministic: z.boolean().default(false),
-    /**
-     * Session scratch retention settings (AC-20).
-     * Controls how long completed session scratch dirs are kept on disk.
-     */
-    session: z
-      .object({
-        /** Days to keep completed session scratch dirs before purging. Default: 7. */
-        retentionDays: z.number().int().min(1).default(7),
-        /** When true and the feature run completes fully, archive scratch to _archive/ instead of deleting. Default: true. */
-        archiveOnFeatureArchive: z.boolean().default(true),
-      })
-      .default(() => ({ retentionDays: 7, archiveOnFeatureArchive: true })),
-    /**
-     * Built-in provider scope configuration (#507).
-     * Controls which working directory each built-in provider uses when
-     * executing git/glob queries in monorepo setups.
-     */
-    providers: z
-      .object({
-        /**
-         * Working directory scope for GitHistoryProvider (#507).
-         * "package" — run git log in packageDir (monorepo-safe default).
-         * "repo" — run git log in repoRoot (full repo history).
-         */
-        historyScope: z.enum(["repo", "package"]).default("package"),
-        /**
-         * Working directory scope for CodeNeighborProvider (#507).
-         * "package" — scan neighbours within packageDir (monorepo-safe default).
-         * "repo" — scan neighbours in repoRoot.
-         */
-        neighborScope: z.enum(["repo", "package"]).default("package"),
-        /**
-         * Cross-package scan depth for CodeNeighborProvider in monorepo mode (#507).
-         * 0 disables cross-package scanning. Default: 1 (one package level up).
-         * Only active when neighborScope is "package" and the story has a workdir.
-         */
-        crossPackageDepth: z.number().int().min(0).default(1),
-      })
-      .default({ historyScope: "package", neighborScope: "package", crossPackageDepth: 1 }),
-    /**
-     * Staleness detection for feature context entries (Amendment A AC-46/AC-47).
-     * Downweights old or contradicted entries in context.md so stale advice
-     * does not crowd out fresh context. No chunks are auto-removed — humans
-     * must edit context.md to remove stale entries.
-     */
-    staleness: z
-      .object({
-        /** Enable staleness detection. Default: true. */
-        enabled: z.boolean().default(true),
-        /**
-         * Number of completed stories after which a context entry is considered
-         * age-stale. Measured by position in context.md (entries are appended
-         * in story order). Default: 10.
-         */
-        maxStoryAge: z.number().int().min(1).default(10),
-        /**
-         * Score multiplier applied to stale chunks (0–1).
-         * 0.4 = stale entry scores 40% of its normal weight, making it less
-         * likely to beat fresh context for the same budget slot.
-         */
-        scoreMultiplier: z.number().min(0).max(1).default(0.4),
-      })
-      .default(() => ({ enabled: true, maxStoryAge: 10, scoreMultiplier: 0.4 })),
-  })
-  .default(() => ({
-    enabled: false,
-    minScore: 0.1,
-    pull: { enabled: false, allowedTools: [], maxCallsPerSession: 5, maxCallsPerRun: 50 },
-    rules: { allowLegacyClaudeMd: false, budgetTokens: 8192 },
-    pluginProviders: [],
-    stages: {},
-    deterministic: false,
-    session: { retentionDays: 7, archiveOnFeatureArchive: true },
-    staleness: { enabled: true, maxStoryAge: 10, scoreMultiplier: 0.4 },
-    providers: { historyScope: "package" as const, neighborScope: "package" as const, crossPackageDepth: 1 },
-  }));
-
-const ContextConfigSchema = z.object({
-  testCoverage: TestCoverageConfigSchema,
-  autoDetect: ContextAutoDetectConfigSchema,
-  fileInjection: z.enum(["keyword", "disabled"]).default("disabled"),
-  featureEngine: FeatureContextEngineConfigSchema.optional(),
-  /** Context Engine settings (Phase 6: enabled by default) */
-  v2: ContextV2ConfigSchema,
-});
-
-const LlmRoutingConfigSchema = z.object({
-  model: ConfiguredModelSchema.optional(),
-  fallbackToKeywords: z.boolean().optional(),
-  cacheDecisions: z.boolean().optional(),
-  mode: z.enum(["one-shot", "per-story", "hybrid"]).optional(),
-  batchMode: z.boolean().optional(), // deprecated, for backward compat
-  timeoutMs: z.number().int().positive({ message: "llm.timeoutMs must be > 0" }).optional(),
-  retries: z.number().int().min(0, { message: "llm.retries must be >= 0" }).optional(),
-  retryDelayMs: z.number().int().min(0, { message: "llm.retryDelayMs must be >= 0" }).optional(),
-});
-
-const RoutingConfigSchema = z.object({
-  strategy: z.enum(["keyword", "llm"]),
-  llm: LlmRoutingConfigSchema.optional(),
-});
-
-const OptimizerConfigSchema = z.object({
-  enabled: z.boolean(),
-  strategy: z.enum(["rule-based", "llm", "noop"]).optional(),
-});
-
-const PluginConfigEntrySchema = z.object({
-  module: z.string().min(1, "plugin.module must be non-empty"),
-  config: z.record(z.string(), z.unknown()).optional(),
-  enabled: z.boolean().default(true),
-});
-
-const HooksConfigSchema = z.object({
-  skipGlobal: z.boolean().optional(),
-  hooks: z.record(z.string(), z.unknown()),
-});
-
-const InteractionConfigSchema = z.object({
-  plugin: z.string().default("cli"),
-  config: z.record(z.string(), z.unknown()).optional(),
-  defaults: z.object({
-    timeout: z.number().int().min(1000).max(3600000).default(600000),
-    fallback: z.enum(["continue", "skip", "escalate", "abort"]).default("escalate"),
-  }),
-  triggers: z
-    .record(
-      z.string(),
-      z.union([
-        z.boolean(),
-        z.object({
-          enabled: z.boolean(),
-          fallback: z.string().optional(),
-          timeout: z.number().optional(),
-        }),
-      ]),
-    )
-    .default({}),
-});
-
-const StorySizeGateConfigSchema = z.object({
-  enabled: z.boolean().default(true),
-  maxAcCount: z.number().int().min(1).max(50).default(10),
-  maxDescriptionLength: z.number().int().min(100).max(10000).default(3000),
-  maxBulletPoints: z.number().int().min(1).max(100).default(12),
-  action: z.enum(["block", "warn", "skip"]).default("block"),
-  maxReplanAttempts: z.number().int().min(1).default(3),
-});
-
-const PromptAuditConfigSchema = z.object({
-  /** When true, every prompt sent to ACP is written to a file for auditing. Default: false. */
-  enabled: z.boolean().default(false),
-  /**
-   * Directory to write audit files into.
-   * Absolute path, or relative to workdir. Defaults to <workdir>/.nax/prompt-audit/ when absent.
-   */
-  dir: z.string().optional(),
-});
-
-const AgentFallbackConfigSchema = z.object({
-  enabled: z.boolean().default(false),
-  map: z.record(z.string().min(1), z.array(z.string().min(1))).default({}),
-  maxHopsPerStory: z.number().int().min(1).max(10).default(2),
-  onQualityFailure: z.boolean().default(false),
-  rebuildContext: z.boolean().default(true),
-});
-
-const AgentAcpConfigSchema = z.object({
-  promptRetries: z.number().int().min(0).max(5).default(0),
-});
-
-const AgentConfigSchema = z.object({
-  protocol: z.literal("acp").default("acp"),
-  default: z.string().trim().min(1, "agent.default must be non-empty").default("claude"),
-  maxInteractionTurns: z.number().int().min(1).max(100).default(20),
-  promptAudit: PromptAuditConfigSchema.default({ enabled: false }),
-  fallback: AgentFallbackConfigSchema.default({
-    enabled: false,
-    map: {},
-    maxHopsPerStory: 2,
-    onQualityFailure: false,
-    rebuildContext: true,
-  }),
-  acp: AgentAcpConfigSchema.default({ promptRetries: 0 }),
-});
-
-const PrecheckConfigSchema = z.object({
-  storySizeGate: StorySizeGateConfigSchema,
-});
-
-export const PromptsConfigSchema = z.object({
-  overrides: z
-    .record(
-      z
-        .string()
-        .refine(
-          (key) => ["no-test", "test-writer", "implementer", "verifier", "single-session", "tdd-simple"].includes(key),
-          {
-            message: "Role must be one of: no-test, test-writer, implementer, verifier, single-session, tdd-simple",
-          },
-        ),
-      z.string().min(1, "Override path must be non-empty"),
-    )
-    .optional(),
-});
-
-const ProjectProfileSchema = z.object({
-  language: z.enum(["typescript", "javascript", "go", "rust", "python", "ruby", "java", "kotlin", "php"]).optional(),
-  type: z.string().optional(),
-  testFramework: z.string().optional(),
-  lintTool: z.string().optional(),
-});
-
-const VALID_AGENT_TYPES = ["claude", "codex", "opencode", "cursor", "windsurf", "aider", "gemini"] as const;
-
-const GenerateConfigSchema = z.object({
-  agents: z.array(z.enum(VALID_AGENT_TYPES)).optional(),
-});
-
-const DebaterPersonaEnum = z.enum(["challenger", "pragmatist", "completionist", "security", "testability"]);
-
-const DebaterSchema = z.object({
-  agent: z.string().min(1, "debater.agent must be non-empty"),
-  model: z.string().min(1, "debater.model must be non-empty").optional(),
-  persona: DebaterPersonaEnum.optional(),
-});
-
-const toObject = (val: unknown): unknown => (val === undefined || val === null ? {} : val);
-
-const RESOLVER_TYPES = ["synthesis", "majority-fail-closed", "majority-fail-open", "custom"] as const;
-
-const makeResolverSchema = (defaultType: (typeof RESOLVER_TYPES)[number]) =>
-  z.preprocess(
-    toObject,
-    z.object({
-      type: z.enum(RESOLVER_TYPES).default(defaultType),
-      agent: z.string().min(1).optional(),
-      model: z.string().min(1).optional(),
-      tieBreaker: z.string().min(1).optional(),
-      maxPromptTokens: z.number().int().positive().optional(),
-    }),
-  );
-
-const DebateStageConfigSchema = (defaults: {
-  enabled: boolean;
-  resolverType: (typeof RESOLVER_TYPES)[number];
-  sessionMode: "one-shot" | "stateful";
-  rounds: number;
-}) =>
-  z.preprocess(
-    toObject,
-    z.object({
-      enabled: z.boolean().default(defaults.enabled),
-      resolver: makeResolverSchema(defaults.resolverType),
-      sessionMode: z.enum(["one-shot", "stateful"]).default(defaults.sessionMode),
-      rounds: z.number().int().min(1).default(defaults.rounds),
-      mode: z.enum(["panel", "hybrid"]).default("panel"),
-      debaters: z.array(DebaterSchema).min(2, "debaters must have at least 2 entries").optional(),
-      timeoutSeconds: z.number().int().positive().default(600),
-      autoPersona: z.boolean().default(false),
-    }),
-  );
-
-const DebateConfigSchema = z.preprocess(
-  toObject,
-  z.object({
-    enabled: z.boolean().default(false),
-    agents: z.number().int().min(2).default(3),
-    maxConcurrentDebaters: z.number().int().min(1).max(10).default(2),
-    stages: z.preprocess(
-      toObject,
-      z.object({
-        plan: DebateStageConfigSchema({ enabled: true, resolverType: "synthesis", sessionMode: "stateful", rounds: 3 }),
-        review: DebateStageConfigSchema({
-          enabled: true,
-          resolverType: "majority-fail-closed",
-          sessionMode: "one-shot",
-          rounds: 2,
-        }),
-        acceptance: DebateStageConfigSchema({
-          enabled: false,
-          resolverType: "majority-fail-closed",
-          sessionMode: "one-shot",
-          rounds: 1,
-        }),
-        rectification: DebateStageConfigSchema({
-          enabled: false,
-          resolverType: "synthesis",
-          sessionMode: "one-shot",
-          rounds: 1,
-        }),
-        escalation: DebateStageConfigSchema({
-          enabled: false,
-          resolverType: "majority-fail-closed",
-          sessionMode: "one-shot",
-          rounds: 1,
-        }),
-      }),
-    ),
-  }),
-);
+// Re-export named schemas consumed by other modules (via config/schema.ts barrel)
+export { AcceptanceConfigSchema } from "./schemas-infra";
+export { AdversarialReviewConfigSchema, ReviewDialogueConfigSchema } from "./schemas-review";
+export { ContextV2ConfigSchema } from "./schemas-context";
+export { PromptsConfigSchema } from "./schemas-infra";
 
 export const NaxConfigSchema = z
   .object({
